CDAP-9165 Add metrics to preview
