+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/preview/DefaultPreviewRunner.java
@@ -17,7 +17,9 @@
 package co.cask.cdap.internal.app.preview;
 
 import co.cask.cdap.api.artifact.ArtifactScope;
+import co.cask.cdap.api.metrics.MetricStore;
 import co.cask.cdap.api.metrics.MetricTimeSeries;
+import co.cask.cdap.api.metrics.MetricsCollectionService;
 import co.cask.cdap.app.preview.DataTracerFactory;
 import co.cask.cdap.app.preview.PreviewRequest;
 import co.cask.cdap.app.preview.PreviewRunner;
@@ -38,6 +40,7 @@ import co.cask.cdap.internal.app.services.ProgramLifecycleService;
 import co.cask.cdap.internal.app.store.RunRecordMeta;
 import co.cask.cdap.logging.appender.LogAppenderInitializer;
 import co.cask.cdap.logging.gateway.handlers.store.ProgramStore;
+import co.cask.cdap.metrics.query.MetricsQueryHelper;
 import co.cask.cdap.proto.BasicThrowable;
 import co.cask.cdap.proto.NamespaceMeta;
 import co.cask.cdap.proto.artifact.AppRequest;
@@ -92,6 +95,8 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
   private final DataTracerFactory dataTracerFactory;
   private final NamespaceAdmin namespaceAdmin;
   private final ProgramStore programStore;
+  private final MetricsCollectionService metricsCollectionService;
+  private final MetricsQueryHelper metricsQueryHelper;
 
   private volatile PreviewStatus status;
   private volatile boolean killedByTimer;
@@ -105,7 +110,8 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
                        SystemArtifactLoader systemArtifactLoader, ProgramRuntimeService programRuntimeService,
                        ProgramLifecycleService programLifecycleService,
                        PreviewStore previewStore, DataTracerFactory dataTracerFactory,
-                       NamespaceAdmin namespaceAdmin, ProgramStore programStore) {
+                       NamespaceAdmin namespaceAdmin, ProgramStore programStore,
+                       MetricsCollectionService metricsCollectionService, MetricsQueryHelper metricsQueryHelper) {
     this.datasetService = datasetService;
     this.logAppenderInitializer = logAppenderInitializer;
     this.applicationLifecycleService = applicationLifecycleService;
@@ -117,6 +123,8 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
     this.dataTracerFactory = dataTracerFactory;
     this.namespaceAdmin = namespaceAdmin;
     this.programStore = programStore;
+    this.metricsCollectionService = metricsCollectionService;
+    this.metricsQueryHelper = metricsQueryHelper;
   }
 
   @Override
@@ -226,11 +234,6 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
   }
 
   @Override
-  public List<MetricTimeSeries> getMetrics() {
-    return new ArrayList<>();
-  }
-
-  @Override
   public ProgramRunId getProgramRunId() {
     return runId;
   }
@@ -238,6 +241,11 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
   @Override
   public RunRecordMeta getRunRecord() {
     return programStore.getRun(programId, runId.getRun());
+  }
+
+  @Override
+  public MetricsQueryHelper getMetricsQueryHelper() {
+    return metricsQueryHelper;
   }
 
   @Override
@@ -255,7 +263,8 @@ public class DefaultPreviewRunner extends AbstractIdleService implements Preview
       applicationLifecycleService.start(),
       systemArtifactLoader.start(),
       programRuntimeService.start(),
-      programLifecycleService.start()
+      programLifecycleService.start(),
+      metricsCollectionService.start()
     ).get();
   }
